@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net5.0;netcoreapp3.1</TargetFrameworks>
+    <TargetFrameworks>net6.0;net5.0;netcoreapp3.1</TargetFrameworks>
     <OutputType>Library</OutputType>
     <RootNamespace>Yarp.ReverseProxy</RootNamespace>
   </PropertyGroup>
@@ -17,7 +17,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Include="..\ReverseProxy.Tests\Utilities\TestResources.cs" Link="Utilities\TestResources.cs" />
+    <Compile Include="..\ReverseProxy.Tests\Common\TestResources.cs" Link="Common\TestResources.cs" />
   </ItemGroup>
 
 </Project>
