@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-#if NET5_0_OR_GREATER
+#if NET
 
 using Yarp.ReverseProxy.Telemetry.Consumption;
 using Prometheus;
