@@ -35,11 +35,11 @@ namespace Yarp.Sample
 
             services.AddHttpContextAccessor();
 
-            // Interface that collects general metrics about the proxy
-            services.AddSingleton<IProxyMetricsConsumer, ProxyMetricsConsumer>();
+            // Interface that collects general metrics about the proxy forwarder
+            services.AddSingleton<IForwarderMetricsConsumer, ForwarderMetricsConsumer>();
 
-            // Registration of a consumer to events for proxy telemetry
-            services.AddTelemetryConsumer<ProxyTelemetryConsumer>();
+            // Registration of a consumer to events for proxy forwarder telemetry
+            services.AddTelemetryConsumer<ForwarderTelemetryConsumer>();
 
             // Registration of a consumer to events for HttpClient telemetry
             // Note: this depends on changes implemented in .NET 5
