@@ -1,10 +1,9 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using Microsoft.Extensions.DependencyInjection;
 using System;
+using Microsoft.Extensions.DependencyInjection;
 using Yarp.ReverseProxy.Configuration;
-using Yarp.ReverseProxy.Service;
 
 namespace Yarp.ReverseProxy.Kubernetes.Protocol
 {
