@@ -4,19 +4,19 @@
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using Yarp.ReverseProxy.Abstractions;
+using Yarp.ReverseProxy.Configuration;
 
 namespace Yarp.ReverseProxy.ServiceFabric
 {
     /// <summary>
     /// Discovers Service Fabric services and builds the corresponding
-    /// <see cref="ProxyRoute"/> and <see cref="Cluster"/> instances that represent them.
+    /// <see cref="RouteConfig"/> and <see cref="ClusterConfig"/> instances that represent them.
     /// </summary>
     internal interface IDiscoverer
     {
         /// <summary>
         /// Execute the discovery and update entities.
         /// </summary>
-        Task<(IReadOnlyList<ProxyRoute> Routes, IReadOnlyList<Cluster> Clusters)> DiscoverAsync(CancellationToken cancellation);
+        Task<(IReadOnlyList<RouteConfig> Routes, IReadOnlyList<ClusterConfig> Clusters)> DiscoverAsync(CancellationToken cancellation);
     }
 }
