@@ -25,7 +25,7 @@ namespace Yarp.ReverseProxy.ServiceFabric
 
         private readonly Cache<IEnumerable<ApplicationWrapper>> _applicationListCache;
         private readonly Cache<IEnumerable<ServiceWrapper>> _serviceListCache;
-        private readonly Cache<IEnumerable<Guid>> _partitionListCache;
+        private readonly Cache<IEnumerable<PartitionWrapper>> _partitionListCache;
         private readonly Cache<IEnumerable<ReplicaWrapper>> _replicaListCache;
         private readonly Cache<string> _serviceManifestCache;
         private readonly Cache<string> _serviceManifestNameCache;
@@ -48,7 +48,7 @@ namespace Yarp.ReverseProxy.ServiceFabric
 
             _applicationListCache = new Cache<IEnumerable<ApplicationWrapper>>(_clock, CacheExpirationOffset);
             _serviceListCache = new Cache<IEnumerable<ServiceWrapper>>(_clock, CacheExpirationOffset);
-            _partitionListCache = new Cache<IEnumerable<Guid>>(_clock, CacheExpirationOffset);
+            _partitionListCache = new Cache<IEnumerable<PartitionWrapper>>(_clock, CacheExpirationOffset);
             _replicaListCache = new Cache<IEnumerable<ReplicaWrapper>>(_clock, CacheExpirationOffset);
             _serviceManifestCache = new Cache<string>(_clock, CacheExpirationOffset);
             _serviceManifestNameCache = new Cache<string>(_clock, CacheExpirationOffset);
@@ -75,7 +75,7 @@ namespace Yarp.ReverseProxy.ServiceFabric
                 cancellation);
         }
 
-        public async Task<IEnumerable<Guid>> GetPartitionListAsync(Uri serviceName, CancellationToken cancellation)
+        public async Task<IEnumerable<PartitionWrapper>> GetPartitionListAsync(Uri serviceName, CancellationToken cancellation)
         {
             return await TryWithCacheFallbackAsync(
                 operationName: "GetPartitionList",
