@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Diagnostics;
 using System.Diagnostics.Tracing;
 using System.Security.Authentication;
@@ -12,7 +13,7 @@ namespace Yarp.ReverseProxy.Telemetry.Consumption
 {
     internal sealed class NetSecurityEventListenerService : EventListenerService<NetSecurityEventListenerService, INetSecurityTelemetryConsumer, INetSecurityMetricsConsumer>
     {
-        private NetSecurityMetrics _previousMetrics;
+        private NetSecurityMetrics? _previousMetrics;
         private NetSecurityMetrics _currentMetrics = new();
         private int _eventCountersCount;
 
@@ -42,7 +43,10 @@ namespace Yarp.ReverseProxy.Telemetry.Consumption
                 return;
             }
 
-            var payload = eventData.Payload;
+#pragma warning disable IDE0007 // Use implicit type
+            // Explicit type here to drop the object? signature of payload elements
+            ReadOnlyCollection<object> payload = eventData.Payload!;
+#pragma warning restore IDE0007 // Use implicit type
 
             switch (eventData.EventId)
             {
@@ -91,8 +95,8 @@ namespace Yarp.ReverseProxy.Telemetry.Consumption
                 return;
             }
 
-            Debug.Assert(eventData.EventName == "EventCounters" && eventData.Payload.Count == 1);
-            var counters = (IDictionary<string, object>)eventData.Payload[0];
+            Debug.Assert(eventData.EventName == "EventCounters" && eventData.Payload!.Count == 1);
+            var counters = (IDictionary<string, object>)eventData.Payload[0]!;
 
             if (!counters.TryGetValue("Mean", out var valueObj))
             {
