@@ -9,7 +9,7 @@ namespace Microsoft.Extensions.DependencyInjection
 {
     public static class TelemetryConsumptionExtensions
     {
-#if NET5_0
+#if NET
         /// <summary>
         /// Registers all telemetry listeners (Proxy, Kestrel, Http, NameResolution, NetSecurity and Sockets).
         /// </summary>
@@ -20,9 +20,9 @@ namespace Microsoft.Extensions.DependencyInjection
 #endif
         public static IServiceCollection AddTelemetryListeners(this IServiceCollection services)
         {
-            services.AddHostedService<ProxyEventListenerService>();
+            services.AddHostedService<ForwarderEventListenerService>();
             services.AddHostedService<KestrelEventListenerService>();
-#if NET5_0
+#if NET
             services.AddHostedService<HttpEventListenerService>();
             services.AddHostedService<NameResolutionEventListenerService>();
             services.AddHostedService<NetSecurityEventListenerService>();
@@ -38,9 +38,9 @@ namespace Microsoft.Extensions.DependencyInjection
         {
             var implementsAny = false;
 
-            if (consumer is IProxyTelemetryConsumer)
+            if (consumer is IForwarderTelemetryConsumer)
             {
-                services.TryAddEnumerable(new ServiceDescriptor(typeof(IProxyTelemetryConsumer), consumer));
+                services.TryAddEnumerable(new ServiceDescriptor(typeof(IForwarderTelemetryConsumer), consumer));
                 implementsAny = true;
             }
 
@@ -50,7 +50,7 @@ namespace Microsoft.Extensions.DependencyInjection
                 implementsAny = true;
             }
 
-#if NET5_0
+#if NET
             if (consumer is IHttpTelemetryConsumer)
             {
                 services.TryAddEnumerable(new ServiceDescriptor(typeof(IHttpTelemetryConsumer), consumer));
@@ -94,9 +94,9 @@ namespace Microsoft.Extensions.DependencyInjection
         {
             var implementsAny = false;
 
-            if (typeof(IProxyTelemetryConsumer).IsAssignableFrom(typeof(TConsumer)))
+            if (typeof(IForwarderTelemetryConsumer).IsAssignableFrom(typeof(TConsumer)))
             {
-                services.AddSingleton(services => (IProxyTelemetryConsumer)services.GetRequiredService<TConsumer>());
+                services.AddSingleton(services => (IForwarderTelemetryConsumer)services.GetRequiredService<TConsumer>());
                 implementsAny = true;
             }
 
@@ -106,7 +106,7 @@ namespace Microsoft.Extensions.DependencyInjection
                 implementsAny = true;
             }
 
-#if NET5_0
+#if NET
             if (typeof(IHttpTelemetryConsumer).IsAssignableFrom(typeof(TConsumer)))
             {
                 services.AddSingleton(services => (IHttpTelemetryConsumer)services.GetRequiredService<TConsumer>());
