@@ -17,7 +17,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
 using Xunit;
 using Yarp.ReverseProxy.Common;
-using Yarp.ReverseProxy.Service.Proxy;
+using Yarp.ReverseProxy.Forwarder;
 using Yarp.ReverseProxy.Telemetry.Consumption;
 
 namespace Yarp.ReverseProxy
@@ -46,9 +46,9 @@ namespace Yarp.ReverseProxy
             else if (approach == RegistrationApproach.Manual)
             {
                 services.AddSingleton<TelemetryConsumer>();
-                services.AddSingleton(services => (IProxyTelemetryConsumer)services.GetRequiredService<TelemetryConsumer>());
+                services.AddSingleton(services => (IForwarderTelemetryConsumer)services.GetRequiredService<TelemetryConsumer>());
                 services.AddSingleton(services => (IKestrelTelemetryConsumer)services.GetRequiredService<TelemetryConsumer>());
-#if NET5_0
+#if NET
                 services.AddSingleton(services => (IHttpTelemetryConsumer)services.GetRequiredService<TelemetryConsumer>());
                 services.AddSingleton(services => (INameResolutionTelemetryConsumer)services.GetRequiredService<TelemetryConsumer>());
                 services.AddSingleton(services => (INetSecurityTelemetryConsumer)services.GetRequiredService<TelemetryConsumer>());
@@ -56,9 +56,9 @@ namespace Yarp.ReverseProxy
 #endif
 
                 services.AddSingleton<SecondTelemetryConsumer>();
-                services.AddSingleton(services => (IProxyTelemetryConsumer)services.GetRequiredService<SecondTelemetryConsumer>());
+                services.AddSingleton(services => (IForwarderTelemetryConsumer)services.GetRequiredService<SecondTelemetryConsumer>());
                 services.AddSingleton(services => (IKestrelTelemetryConsumer)services.GetRequiredService<SecondTelemetryConsumer>());
-#if NET5_0
+#if NET
                 services.AddSingleton(services => (IHttpTelemetryConsumer)services.GetRequiredService<SecondTelemetryConsumer>());
                 services.AddSingleton(services => (INameResolutionTelemetryConsumer)services.GetRequiredService<SecondTelemetryConsumer>());
                 services.AddSingleton(services => (INetSecurityTelemetryConsumer)services.GetRequiredService<SecondTelemetryConsumer>());
@@ -102,10 +102,10 @@ namespace Yarp.ReverseProxy
             var expected = new[]
             {
                 "OnRequestStart-Kestrel",
-                "OnProxyInvoke",
-                "OnProxyStart",
-                "OnProxyStage-SendAsyncStart",
-#if NET5_0
+                "OnForwarderInvoke",
+                "OnForwarderStart",
+                "OnForwarderStage-SendAsyncStart",
+#if NET
                 "OnRequestStart",
                 "OnConnectStart",
                 "OnConnectStop",
@@ -118,10 +118,10 @@ namespace Yarp.ReverseProxy
                 "OnResponseHeadersStop",
                 "OnRequestStop",
 #endif
-                "OnProxyStage-SendAsyncStop",
-                "OnProxyStage-ResponseContentTransferStart",
+                "OnForwarderStage-SendAsyncStop",
+                "OnForwarderStage-ResponseContentTransferStart",
                 "OnContentTransferred",
-                "OnProxyStop",
+                "OnForwarderStop",
                 "OnRequestStop-Kestrel"
             };
 
@@ -132,7 +132,7 @@ namespace Yarp.ReverseProxy
             }
         }
 
-#if NET5_0
+#if NET
         [Theory]
         [InlineData(RegistrationApproach.WithInstanceHelper)]
         [InlineData(RegistrationApproach.WithGenericHelper)]
@@ -177,9 +177,9 @@ namespace Yarp.ReverseProxy
         private class SecondTelemetryConsumer : TelemetryConsumer { }
 
         private class TelemetryConsumer :
-            IProxyTelemetryConsumer,
+            IForwarderTelemetryConsumer,
             IKestrelTelemetryConsumer
-#if NET5_0
+#if NET
             ,
             IHttpTelemetryConsumer,
             INameResolutionTelemetryConsumer,
@@ -202,18 +202,18 @@ namespace Yarp.ReverseProxy
                 }
             }
 
-            public void OnProxyStart(DateTime timestamp, string destinationPrefix) => AddStage(nameof(OnProxyStart), timestamp);
-            public void OnProxyStop(DateTime timestamp, int statusCode) => AddStage(nameof(OnProxyStop), timestamp);
-            public void OnProxyFailed(DateTime timestamp, ProxyError error) => AddStage(nameof(OnProxyFailed), timestamp);
-            public void OnProxyStage(DateTime timestamp, Telemetry.Consumption.ProxyStage stage) => AddStage($"{nameof(OnProxyStage)}-{stage}", timestamp);
+            public void OnForwarderStart(DateTime timestamp, string destinationPrefix) => AddStage(nameof(OnForwarderStart), timestamp);
+            public void OnForwarderStop(DateTime timestamp, int statusCode) => AddStage(nameof(OnForwarderStop), timestamp);
+            public void OnForwarderFailed(DateTime timestamp, ForwarderError error) => AddStage(nameof(OnForwarderFailed), timestamp);
+            public void OnForwarderStage(DateTime timestamp, Telemetry.Consumption.ForwarderStage stage) => AddStage($"{nameof(OnForwarderStage)}-{stage}", timestamp);
             public void OnContentTransferring(DateTime timestamp, bool isRequest, long contentLength, long iops, TimeSpan readTime, TimeSpan writeTime) => AddStage(nameof(OnContentTransferring), timestamp);
             public void OnContentTransferred(DateTime timestamp, bool isRequest, long contentLength, long iops, TimeSpan readTime, TimeSpan writeTime, TimeSpan firstReadTime) => AddStage(nameof(OnContentTransferred), timestamp);
-            public void OnProxyInvoke(DateTime timestamp, string clusterId, string routeId, string destinationId)
+            public void OnForwarderInvoke(DateTime timestamp, string clusterId, string routeId, string destinationId)
             {
-                AddStage(nameof(OnProxyInvoke), timestamp);
+                AddStage(nameof(OnForwarderInvoke), timestamp);
                 PerClusterTelemetry.TryAdd((clusterId, GetType()), _stages.Value);
             }
-#if NET5_0
+#if NET
             public void OnRequestStart(DateTime timestamp, string scheme, string host, int port, string pathAndQuery, int versionMajor, int versionMinor, HttpVersionPolicy versionPolicy)
             {
                 AddStage(nameof(OnRequestStart), timestamp);
@@ -262,8 +262,8 @@ namespace Yarp.ReverseProxy
                 {
                     var services = proxyBuilder.Services;
 
-                    services.AddSingleton<IProxyMetricsConsumer>(consumer);
-#if NET5_0
+                    services.AddSingleton<IForwarderMetricsConsumer>(consumer);
+#if NET
                     services.AddSingleton<IKestrelMetricsConsumer>(consumer);
                     services.AddSingleton<IHttpMetricsConsumer>(consumer);
                     services.AddSingleton<ISocketsMetricsConsumer>(consumer);
@@ -289,8 +289,8 @@ namespace Yarp.ReverseProxy
                 catch { }
 
                 await Task.WhenAll(
-                    WaitAsync(() => consumer.ProxyMetrics.LastOrDefault()?.RequestsStarted > 0, nameof(ProxyMetrics))
-#if NET5_0
+                    WaitAsync(() => consumer.ProxyMetrics.LastOrDefault()?.RequestsStarted > 0, nameof(ForwarderMetrics))
+#if NET
                     ,
                     WaitAsync(() => consumer.KestrelMetrics.LastOrDefault()?.TotalConnections > 0, nameof(KestrelMetrics)),
                     WaitAsync(() => consumer.HttpMetrics.LastOrDefault()?.RequestsStarted > 0, nameof(HttpMetrics)),
@@ -302,7 +302,7 @@ namespace Yarp.ReverseProxy
             });
 
             VerifyTimestamp(consumer.ProxyMetrics.Last().Timestamp);
-#if NET5_0
+#if NET
             VerifyTimestamp(consumer.KestrelMetrics.Last().Timestamp);
             VerifyTimestamp(consumer.HttpMetrics.Last().Timestamp);
             VerifyTimestamp(consumer.SocketsMetrics.Last().Timestamp);
@@ -331,8 +331,8 @@ namespace Yarp.ReverseProxy
         }
 
         private sealed class MetricsConsumer :
-            IProxyMetricsConsumer
-#if NET5_0
+            IForwarderMetricsConsumer
+#if NET
             ,
             IKestrelMetricsConsumer,
             IHttpMetricsConsumer,
@@ -341,8 +341,8 @@ namespace Yarp.ReverseProxy
             ISocketsMetricsConsumer
 #endif
         {
-            public readonly ConcurrentQueue<ProxyMetrics> ProxyMetrics = new ConcurrentQueue<ProxyMetrics>();
-#if NET5_0
+            public readonly ConcurrentQueue<ForwarderMetrics> ProxyMetrics = new ConcurrentQueue<ForwarderMetrics>();
+#if NET
             public readonly ConcurrentQueue<KestrelMetrics> KestrelMetrics = new();
             public readonly ConcurrentQueue<HttpMetrics> HttpMetrics = new();
             public readonly ConcurrentQueue<SocketsMetrics> SocketsMetrics = new();
@@ -350,8 +350,8 @@ namespace Yarp.ReverseProxy
             public readonly ConcurrentQueue<NameResolutionMetrics> NameResolutionMetrics = new();
 #endif
 
-            public void OnProxyMetrics(ProxyMetrics oldMetrics, ProxyMetrics newMetrics) => ProxyMetrics.Enqueue(newMetrics);
-#if NET5_0
+            public void OnForwarderMetrics(ForwarderMetrics oldMetrics, ForwarderMetrics newMetrics) => ProxyMetrics.Enqueue(newMetrics);
+#if NET
             public void OnKestrelMetrics(KestrelMetrics oldMetrics, KestrelMetrics newMetrics) => KestrelMetrics.Enqueue(newMetrics);
             public void OnSocketsMetrics(SocketsMetrics oldMetrics, SocketsMetrics newMetrics) => SocketsMetrics.Enqueue(newMetrics);
             public void OnNetSecurityMetrics(NetSecurityMetrics oldMetrics, NetSecurityMetrics newMetrics) => NetSecurityMetrics.Enqueue(newMetrics);
