@@ -4,17 +4,15 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
 using System.Text.Json;
+using System.Threading.Tasks;
 using k8s;
 using k8s.Models;
 using Newtonsoft.Json.Linq;
-using YamlDotNet.Serialization;
+using Xunit;
 using Yarp.ReverseProxy.Kubernetes.Controller.Caching;
 using Yarp.ReverseProxy.Kubernetes.Controller.Converters;
 using Yarp.ReverseProxy.Kubernetes.Controller.Services;
-using Xunit;
 
 namespace IngressController.Tests
 {
