@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Fabric.Health;
 using System.Fabric.Query;
 using Microsoft.ServiceFabric.Services.Communication;
-using Yarp.ReverseProxy.Abstractions;
+using Yarp.ReverseProxy.Configuration;
 
 namespace Yarp.ReverseProxy.ServiceFabric.Tests
 {
@@ -35,9 +35,13 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
                 ServiceKind = serviceKind,
             };
         }
-        internal static Guid FakePartition()
+        internal static PartitionWrapper FakePartition()
         {
-            return Guid.NewGuid();
+            return new PartitionWrapper
+            {
+                Id = Guid.NewGuid(),
+                Name = "Test"
+            };
         }
         internal static ReplicaWrapper FakeReplica(Uri serviceName, int id)
         {
@@ -70,12 +74,12 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
         }
 
         /// <summary>
-        /// Build a <see cref="Destination" /> from a Service Fabric <see cref="ReplicaWrapper" />.
+        /// Build a <see cref="DestinationConfig" /> from a Service Fabric <see cref="ReplicaWrapper" />.
         /// </summary>
         /// <remarks>
         /// The address JSON of the replica is expected to have exactly one endpoint, and that one will be used.
         /// </remarks>
-        internal static KeyValuePair<string, Destination> BuildDestinationFromReplica(ReplicaWrapper replica, string healthListenerName = null)
+        internal static KeyValuePair<string, DestinationConfig> BuildDestinationFromReplicaAndPartition(ReplicaWrapper replica, PartitionWrapper partition, string healthListenerName = null)
         {
             ServiceEndpointCollection.TryParseEndpointsString(replica.ReplicaAddress, out var endpoints);
             endpoints.TryGetFirstEndpointAddress(out var address);
@@ -86,13 +90,20 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
                 endpoints.TryGetEndpointAddress(healthListenerName, out healthAddressUri);
             }
 
+            var destinationId = $"{partition.Id}/{replica.Id}";
+
             return KeyValuePair.Create(
-                replica.Id.ToString(),
-                new Destination
+                destinationId,
+                new DestinationConfig
                 {
                     Address = address,
                     Health = healthAddressUri,
-                    Metadata = null,
+                    Metadata = new Dictionary<string, string>
+                {
+                    { "PartitionId", partition.Id.ToString() ?? string.Empty },
+                    { "NamedPartitionName", partition.Name ?? string.Empty },
+                    { "ReplicaId", replica.Id.ToString() ?? string.Empty }
+                }
                 });
         }
     }
