@@ -4,12 +4,12 @@
 using System;
 using System.Collections.Generic;
 using System.Security.Authentication;
-using System.Text;
 using FluentAssertions;
 using Xunit;
-using Yarp.ReverseProxy.Abstractions;
-using Yarp.ReverseProxy.Abstractions.ClusterDiscovery.Contract;
-using Yarp.ReverseProxy.Service.Proxy;
+using Yarp.ReverseProxy.Configuration;
+using Yarp.ReverseProxy.LoadBalancing;
+using Yarp.ReverseProxy.Forwarder;
+using Yarp.ReverseProxy.SessionAffinity;
 
 namespace Yarp.ReverseProxy.ServiceFabric.Tests
 {
@@ -26,11 +26,19 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
                 { "YARP.Backend.BackendId", "MyCoolClusterId" },
                 { "YARP.Backend.LoadBalancingPolicy", "LeastRequests" },
                 { "YARP.Backend.SessionAffinity.Enabled", "true" },
-                { "YARP.Backend.SessionAffinity.Mode", "Cookie" },
+                { "YARP.Backend.SessionAffinity.Policy", "Cookie" },
                 { "YARP.Backend.SessionAffinity.FailurePolicy", "Return503Error" },
-                { "YARP.Backend.SessionAffinity.Settings.ParameterA", "ValueA" },
-                { "YARP.Backend.SessionAffinity.Settings.ParameterB", "ValueB" },
+                { "YARP.Backend.SessionAffinity.AffinityKeyName", "Key1" },
+                { "YARP.Backend.SessionAffinity.Cookie.Domain", "localhost" },
+                { "YARP.Backend.SessionAffinity.Cookie.Expiration", "03:00:00" },
+                { "YARP.Backend.SessionAffinity.Cookie.HttpOnly", "true" },
+                { "YARP.Backend.SessionAffinity.Cookie.IsEssential", "true" },
+                { "YARP.Backend.SessionAffinity.Cookie.MaxAge", "1.00:00:00" },
+                { "YARP.Backend.SessionAffinity.Cookie.Path", "mypath" },
+                { "YARP.Backend.SessionAffinity.Cookie.SameSite", "Strict" },
+                { "YARP.Backend.SessionAffinity.Cookie.SecurePolicy", "SameAsRequest" },
                 { "YARP.Backend.HttpRequest.Timeout", "00:00:17" },
+                { "YARP.Backend.HttpRequest.AllowResponseBuffering", "true" },
                 { "YARP.Backend.HttpRequest.Version", "1.1" },
 #if NET
                 { "YARP.Backend.HttpRequest.VersionPolicy", "RequestVersionExact" },
@@ -60,32 +68,40 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
 
             var cluster = LabelsParser.BuildCluster(_testServiceName, labels, null);
 
-            var expectedCluster = new Cluster
+            var expectedCluster = new ClusterConfig
             {
-                Id = "MyCoolClusterId",
+                ClusterId = "MyCoolClusterId",
                 LoadBalancingPolicy = LoadBalancingPolicies.LeastRequests,
-                SessionAffinity = new SessionAffinityOptions
+                SessionAffinity = new SessionAffinityConfig
                 {
                     Enabled = true,
-                    Mode = SessionAffinityConstants.Modes.Cookie,
-                    FailurePolicy = SessionAffinityConstants.AffinityFailurePolicies.Return503Error,
-                    Settings = new Dictionary<string, string>
+                    Policy = SessionAffinityConstants.Policies.Cookie,
+                    FailurePolicy = SessionAffinityConstants.FailurePolicies.Return503Error,
+                    AffinityKeyName = "Key1",
+                    Cookie = new SessionAffinityCookieConfig
                     {
-                        { "ParameterA", "ValueA" },
-                        { "ParameterB", "ValueB" }
+                        Domain = "localhost",
+                        Expiration = TimeSpan.FromHours(3),
+                        HttpOnly = true,
+                        IsEssential = true,
+                        MaxAge = TimeSpan.FromDays(1),
+                        Path = "mypath",
+                        SameSite = Microsoft.AspNetCore.Http.SameSiteMode.Strict,
+                        SecurePolicy = Microsoft.AspNetCore.Http.CookieSecurePolicy.SameAsRequest
                     }
                 },
-                HttpRequest = new RequestProxyOptions
+                HttpRequest = new ForwarderRequestConfig
                 {
                     Timeout = TimeSpan.FromSeconds(17),
                     Version = new Version(1, 1),
+                    AllowResponseBuffering = true,
 #if NET
                     VersionPolicy = System.Net.Http.HttpVersionPolicy.RequestVersionExact
 #endif
                 },
-                HealthCheck = new HealthCheckOptions
+                HealthCheck = new HealthCheckConfig
                 {
-                    Active = new ActiveHealthCheckOptions
+                    Active = new ActiveHealthCheckConfig
                     {
                         Enabled = true,
                         Interval = TimeSpan.FromSeconds(5),
@@ -93,7 +109,7 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
                         Path = "/api/health",
                         Policy = "MyActiveHealthPolicy"
                     },
-                    Passive = new PassiveHealthCheckOptions
+                    Passive = new PassiveHealthCheckConfig
                     {
                         Enabled = true,
                         Policy = "MyPassiveHealthPolicy",
@@ -104,17 +120,17 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
                 {
                     { "Foo", "Bar" },
                 },
-                HttpClient = new ProxyHttpClientOptions
+                HttpClient = new HttpClientConfig
                 {
                     ActivityContextHeaders = ActivityContextHeaders.BaggageAndCorrelationContext,
                     DangerousAcceptAnyServerCertificate = true,
 #if NET
                     EnableMultipleHttp2Connections = false,
-                    RequestHeaderEncoding = Encoding.GetEncoding("utf-8"),
+                    RequestHeaderEncoding = "utf-8",
 #endif
                     MaxConnectionsPerServer = 1000,
                     SslProtocols = SslProtocols.Tls12,
-                    WebProxy = new WebProxyOptions
+                    WebProxy = new WebProxyConfig
                     {
                         Address = new Uri("https://10.20.30.40"),
                         BypassOnLocal = true,
@@ -131,33 +147,29 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
             var labels = new Dictionary<string, string>()
             {
                 { "YARP.Backend.BackendId", "MyCoolClusterId" },
+                { "YARP.Backend.SessionAffinity.AffinityKeyName", "Key1" }
             };
 
             var cluster = LabelsParser.BuildCluster(_testServiceName, labels, null);
 
-            var expectedCluster = new Cluster
+            var expectedCluster = new ClusterConfig
             {
-                Id = "MyCoolClusterId",
-                SessionAffinity = new SessionAffinityOptions
+                ClusterId = "MyCoolClusterId",
+                SessionAffinity = new SessionAffinityConfig
                 {
-                    Enabled = false,
+                    AffinityKeyName = "Key1",
+                    Cookie = new SessionAffinityCookieConfig()
                 },
-                HttpRequest = new RequestProxyOptions(),
-                HealthCheck = new HealthCheckOptions
+                HttpRequest = new ForwarderRequestConfig(),
+                HealthCheck = new HealthCheckConfig
                 {
-                    Active = new ActiveHealthCheckOptions
-                    {
-                        Enabled = false,
-                    },
-                    Passive = new PassiveHealthCheckOptions
-                    {
-                        Enabled = false,
-                    }
+                    Active = new ActiveHealthCheckConfig(),
+                    Passive = new PassiveHealthCheckConfig()
                 },
                 Metadata = new Dictionary<string, string>(),
-                HttpClient = new ProxyHttpClientOptions
+                HttpClient = new HttpClientConfig
                 {
-                    WebProxy = new WebProxyOptions
+                    WebProxy = new WebProxyConfig
                     {
                     }
                 }
@@ -172,7 +184,9 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
         [InlineData("false", false)]
         [InlineData("False", false)]
         [InlineData("FALSE", false)]
-        public void BuildCluster_HealthCheckOptions_Enabled_Valid(string label, bool expected)
+        [InlineData(null, null)]
+        [InlineData("", null)]
+        public void BuildCluster_HealthCheckOptions_Enabled_Valid(string label, bool? expected)
         {
             var labels = new Dictionary<string, string>()
             {
@@ -186,8 +200,8 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
         }
 
         [Theory]
-        [InlineData(null)]
-        [InlineData("")]
+        [InlineData("notbool")]
+        [InlineData(" ")]
         public void BuildCluster_HealthCheckOptions_Enabled_Invalid(string label)
         {
             var labels = new Dictionary<string, string>()
@@ -215,7 +229,7 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
 
             var cluster = LabelsParser.BuildCluster(_testServiceName, labels, null);
 
-            cluster.Id.Should().Be(_testServiceName.ToString());
+            cluster.ClusterId.Should().Be(_testServiceName.ToString());
         }
 
         [Theory]
@@ -244,7 +258,7 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
                 { key, invalidValue },
             };
 
-            Func<Cluster> func = () => LabelsParser.BuildCluster(_testServiceName, labels, null);
+            Func<ClusterConfig> func = () => LabelsParser.BuildCluster(_testServiceName, labels, null);
 
             func.Should().Throw<ConfigException>().WithMessage($"Could not convert label {key}='{invalidValue}' *");
         }
@@ -275,9 +289,9 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
 
             var routes = LabelsParser.BuildRoutes(_testServiceName, labels);
 
-            var expectedRoutes = new List<ProxyRoute>
+            var expectedRoutes = new List<RouteConfig>
             {
-                new ProxyRoute
+                new RouteConfig
                 {
                     RouteId = "MyCoolClusterId:MyRoute",
                     Match = new RouteMatch
@@ -338,9 +352,9 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
 
             var routes = LabelsParser.BuildRoutes(_testServiceName, labels);
 
-            var expectedRoutes = new List<ProxyRoute>
+            var expectedRoutes = new List<RouteConfig>
             {
-                new ProxyRoute
+                new RouteConfig
                 {
                     RouteId = "MyCoolClusterId:MyRoute",
                     Match = new RouteMatch
@@ -368,9 +382,9 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
 
             var routes = LabelsParser.BuildRoutes(_testServiceName, labels);
 
-            var expectedRoutes = new List<ProxyRoute>
+            var expectedRoutes = new List<RouteConfig>
             {
-                new ProxyRoute
+                new RouteConfig
                 {
                     RouteId = "MyCoolClusterId:MyRoute",
                     Match = new RouteMatch
@@ -402,9 +416,9 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
 
             var routes = LabelsParser.BuildRoutes(_testServiceName, labels);
 
-            var expectedRoutes = new List<ProxyRoute>
+            var expectedRoutes = new List<RouteConfig>
             {
-                new ProxyRoute
+                new RouteConfig
                 {
                     RouteId = $"{Uri.EscapeDataString(_testServiceName.ToString())}:MyRoute",
                     Match = new RouteMatch
@@ -429,9 +443,9 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
 
             var routes = LabelsParser.BuildRoutes(_testServiceName, labels);
 
-            var expectedRoutes = new List<ProxyRoute>
+            var expectedRoutes = new List<RouteConfig>
             {
-                new ProxyRoute
+                new RouteConfig
                 {
                     RouteId = $"{Uri.EscapeDataString(_testServiceName.ToString())}:MyRoute",
                     Match = new RouteMatch
@@ -455,7 +469,7 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
                 { "YARP.Routes.MyRoute.Order", "this is no number" },
             };
 
-            Func<List<ProxyRoute>> func = () => LabelsParser.BuildRoutes(_testServiceName, labels);
+            Func<List<RouteConfig>> func = () => LabelsParser.BuildRoutes(_testServiceName, labels);
 
             func.Should()
                 .Throw<ConfigException>()
@@ -479,9 +493,9 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
 
             var routes = LabelsParser.BuildRoutes(_testServiceName, labels);
 
-            var expectedRoutes = new List<ProxyRoute>
+            var expectedRoutes = new List<RouteConfig>
             {
-                new ProxyRoute
+                new RouteConfig
                 {
                     RouteId = $"MyCoolClusterId:{routeName}",
                     Match = new RouteMatch
@@ -517,7 +531,7 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
             };
             labels[invalidKey] = value;
 
-            Func<List<ProxyRoute>> func = () => LabelsParser.BuildRoutes(_testServiceName, labels);
+            Func<List<RouteConfig>> func = () => LabelsParser.BuildRoutes(_testServiceName, labels);
 
             func.Should()
                 .Throw<ConfigException>()
@@ -539,7 +553,7 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
             };
             labels[invalidKey] = value;
 
-            Func<List<ProxyRoute>> func = () => LabelsParser.BuildRoutes(_testServiceName, labels);
+            Func<List<RouteConfig>> func = () => LabelsParser.BuildRoutes(_testServiceName, labels);
 
             func.Should()
                 .Throw<ConfigException>()
@@ -563,7 +577,7 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
             labels[invalidKey] = value;
 
             // Act
-            Func<List<ProxyRoute>> func = () => LabelsParser.BuildRoutes(_testServiceName, labels);
+            Func<List<RouteConfig>> func = () => LabelsParser.BuildRoutes(_testServiceName, labels);
 
             // Assert
             func.Should()
@@ -586,7 +600,7 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
             labels[invalidKey] = value;
 
             // Act
-            Func<List<ProxyRoute>> func = () => LabelsParser.BuildRoutes(_testServiceName, labels);
+            Func<List<RouteConfig>> func = () => LabelsParser.BuildRoutes(_testServiceName, labels);
 
             // Assert
             func.Should()
@@ -611,9 +625,9 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
 
             var routes = LabelsParser.BuildRoutes(_testServiceName, labels);
 
-            var expectedRoutes = new List<ProxyRoute>
+            var expectedRoutes = new List<RouteConfig>
             {
-                new ProxyRoute
+                new RouteConfig
                 {
                     RouteId = $"MyCoolClusterId:MyRoute0",
                     Match = new RouteMatch
@@ -661,9 +675,9 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
 
             var routes = LabelsParser.BuildRoutes(_testServiceName, labels);
 
-            var expectedRoutes = new List<ProxyRoute>
+            var expectedRoutes = new List<RouteConfig>
             {
-                new ProxyRoute
+                new RouteConfig
                 {
                     RouteId = "MyCoolClusterId:MyRoute",
                     Match = new RouteMatch
@@ -699,9 +713,9 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
 
             var routes = LabelsParser.BuildRoutes(_testServiceName, labels);
 
-            var expectedRoutes = new List<ProxyRoute>
+            var expectedRoutes = new List<RouteConfig>
             {
-                new ProxyRoute
+                new RouteConfig
                 {
                     RouteId = "MyCoolClusterId:MyRoute",
                     Match = new RouteMatch
@@ -713,7 +727,7 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
                     ClusterId = "MyCoolClusterId",
                     Metadata = new Dictionary<string, string> { { "Foo", "Bar" } },
                 },
-                new ProxyRoute
+                new RouteConfig
                 {
                     RouteId = "MyCoolClusterId:CoolRoute",
                     Match = new RouteMatch
@@ -724,7 +738,7 @@ namespace Yarp.ReverseProxy.ServiceFabric.Tests
                     ClusterId = "MyCoolClusterId",
                     Metadata = new Dictionary<string, string>(),
                 },
-                new ProxyRoute
+                new RouteConfig
                 {
                     RouteId = "MyCoolClusterId:EvenCoolerRoute",
                     Match = new RouteMatch
