@@ -4,7 +4,7 @@
 using System;
 using Xunit;
 
-namespace Yarp.ReverseProxy.Utilities
+namespace Yarp.ReverseProxy.Utilities.Tests
 {
     public class RandomFactoryTests
     {
